+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Inspector/PlatformGameObjectRemoverEditor.cs
@@ -24,13 +24,12 @@ namespace KRT.VRCQuestTools.Inspector
 #endif
             EditorGUILayout.HelpBox(i18n.PlatformGameObjectRemoverEditorDescription, MessageType.Info);
 
+            EditorGUILayout.HelpBox(i18n.PlatformTargetSettingsIsRequiredToEnforcePlatform, MessageType.Info);
+
             EditorGUILayout.Space();
 
             var so = new SerializedObject(target);
             so.Update();
-
-            var buildTarget = so.FindProperty("buildTarget");
-            EditorGUILayout.PropertyField(buildTarget, new GUIContent(i18n.BuildTargetLabel, i18n.BuildTargetTooltip));
 
             var removeOnPC = so.FindProperty("removeOnPC");
             EditorGUILayout.PropertyField(removeOnPC, new GUIContent(i18n.PlatformGameObjectRemoverEditorRemoveOnPCLabel));