+++ Packages/com.github.kurotu.vrc-quest-tools/Runtime/Components/VRCQuestToolsPlatformComponent.cs
@@ -1,15 +1,1 @@
-using KRT.VRCQuestTools.Models;
-
-namespace KRT.VRCQuestTools.Components
-{
-    /// <summary>
-    /// Editor only component which depends on build target.
-    /// </summary>
-    public class VRCQuestToolsPlatformComponent : VRCQuestToolsEditorOnly
-    {
-        /// <summary>
-        /// Build target to control this component.
-        /// </summary>
-        public BuildTarget buildTarget = BuildTarget.Auto;
-    }
-}
+// removed source code.